Fix contact page for hyrax
