--- app/controllers/contact_form_controller.rb
@@ -1,3 +1,0 @@
-class ContactFormController < ApplicationController
-  include Hyrax::ContactFormControllerBehavior
-end
