+++ app/controllers/hyrax/contact_form_controller.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+module Hyrax
+  class ContactFormController < ApplicationController
+    include ContactFormControllerBehavior
+  end
+end
