--- app/views/contact_form/create.html.erb
@@ -1,1 +1,0 @@
-<h1>Contact Form</h1>
